+++ core/src/main/java/com/aemproof/core/models/package-info.java
@@ -1,19 +1,6 @@
-/*
- *  Copyright 2015 Adobe Systems Incorporated
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+/**
+ * Sling Models: the view layer's contract. HTL scripts only ever talk to the interfaces in this
+ * package; implementations live in {@code impl} and are registered by resource type.
  */
-@Version("1.0")
+@org.osgi.annotation.versioning.Version("1.0.0")
 package com.aemproof.core.models;
-
-import org.osgi.annotation.versioning.Version;